requirements.txt
mysql-connector-python
Flask
pytest
